+++ crates/epaint/src/shapes/mod.rs
@@ -1,6 +1,7 @@
 mod bezier;
 mod circle;
 mod ellipse;
+mod paint_callback;
 mod path;
 mod rect;
 mod shape;
@@ -10,10 +11,9 @@ pub use self::{
     bezier::{CubicBezierShape, QuadraticBezierShape},
     circle::CircleShape,
     ellipse::EllipseShape,
+    paint_callback::{PaintCallback, PaintCallbackInfo},
     path::PathShape,
     rect::RectShape,
-    shape::PaintCallback,
-    shape::PaintCallbackInfo,
     shape::Shape,
     text::TextShape,
 };
